Add support for separating the channels of an input device into separate audio inputs, including cosine volume panning to get the volume right.

+++ NowSoundLib/NowSoundTrack.cpp
@@ -153,7 +153,7 @@ namespace NowSound
         _debugLog{},
         _requiredSamplesHistogram { MagicNumbers::AudioQuantumHistogramCapacity },
 		_sinceLastSampleTimingHistogram{ MagicNumbers::AudioQuantumHistogramCapacity },
-		_recentVolumeHistogram{ Clock::Instance().TimeToSamples(MagicNumbers::RecentVolumeDuration).Value() }
+		_recentVolumeHistogram{ (int)Clock::Instance().TimeToSamples(MagicNumbers::RecentVolumeDuration).Value() }
 	{
         Check(_lastSampleTime.Value() >= 0);
